Slice download client adapters into workflows
