+++ BACKEND_ARCHITECTURE.md
@@ -93,6 +93,27 @@ Background workers expose DI-facing processor contracts for deterministic cycle/
 | `FfmpegInstallBackgroundService` | `FfmpegInstallProcessor` | Owns non-blocking ffprobe/ffmpeg availability checks and install attempts. | Must not block host startup or mutate unrelated settings. | Runs once outside request startup; failures are reported without stopping the host. | Rechecks installed binaries before downloading/installing. | `DownloadHub` receives `FfmpegInstallStatus`. |
 | `UnmatchedScanBackgroundService` | `UnmatchedScanProcessor` | Owns Library Import unmatched-file scan job status `Queued -> Processing -> Completed/Failed` and cached result replacement. | Must not import matched files or create audiobook records. | Queue-driven; failed/finished jobs can be superseded by a new explicit scan. | Groups files deterministically and clears stale unmatched results for the scanned root. | `SettingsHub` receives `UnmatchedScanComplete`. |
 
+## Download Client Adapter Slicing
+
+Each concrete download-client adapter should be a thin facade over client-specific workflows. This keeps the application-facing `IDownloadClientAdapter` contract stable while preventing one large adapter class from owning HTTP/XML-RPC calls, response parsing, monitor polling policy, import path resolution, and client cleanup at the same time.
+
+Use this shape for every supported client:
+
+- `<Client>Adapter` exposes `IDownloadClientAdapter`, `ClientId`, `ClientType`, and `Protocol`, then delegates.
+- `<Client>ConnectionTester` owns connection-test behavior and user-facing connection messages.
+- `<Client>AddWorkflow` owns submission behavior and client-specific add quirks.
+- `<Client>QueueFetchWorkflow` owns live queue polling, ID-filtered monitor polling, and monitor/display failure semantics.
+- `<Client>ItemFetchWorkflow` owns full external item-list display behavior.
+- `<Client>HistoryFetchWorkflow` or `<Client>HistoryEnrichmentWorkflow` owns history lookup/enrichment where the client exposes history.
+- `<Client>RemovalWorkflow` owns external client cleanup only.
+- `<Client>ImportItemResolver` owns import-source resolution and retry-aware path lookup.
+- `<Client>ResponseMapper` should stay focused on translating client payloads into Listenarr models.
+- `<Client>RequestPlanner`, `<Client>RpcClient`, and auth/session helpers own protocol mechanics.
+
+Do not move concrete downloader behavior into `listenarr.application`. Client-specific HTTP, XML-RPC, JSON parsing, history quirks, category behavior, seed-limit evaluation, and cleanup semantics belong in `listenarr.infrastructure/DownloadClients/<Client>`. Do not introduce a shared abstract base adapter for all clients unless the behavior is genuinely identical. Prefer uniform class roles over forced inheritance because NZBGet, SABnzbd, qBittorrent, and Transmission expose different APIs and different lifecycle semantics.
+
+When refactoring a client, preserve unique behavior explicitly in the workflow that owns it. Examples include NZBGet history/final-path enrichment, SABnzbd queue/history reconciliation, qBittorrent tracker injection and post-import category marking, and Transmission local ID filtering for monitor polls.
+
 The main download handoff is:
 
 1. `DownloadMonitorService` observes the external client and persists `Completed`.
