+++ listenarr.infrastructure/DownloadClients/Qbittorrent/QbittorrentAdapter.cs
@@ -15,7 +15,6 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program. If not, see <https://www.gnu.org/licenses/>.
  */
-using System.Net;
 using System.Text.Json;
 using Listenarr.Infrastructure.Torrents;
 using Microsoft.Extensions.Logging;
@@ -23,7 +22,9 @@ using Microsoft.Extensions.Logging;
 namespace Listenarr.Infrastructure.DownloadClients.Qbittorrent
 {
     /// <summary>
-    /// qBittorrent protocol implementation.
+    /// qBittorrent protocol facade.
+    /// Keep client-specific workflows behind this adapter so every supported
+    /// download client exposes the same thin IDownloadClientAdapter surface.
     /// </summary>
     public class QbittorrentAdapter : IDownloadClientAdapter
     {
@@ -31,417 +32,84 @@ namespace Listenarr.Infrastructure.DownloadClients.Qbittorrent
         public string ClientType => "qbittorrent";
         public DownloadProtocol Protocol => DownloadProtocol.Torrent;
 
-        private readonly IHttpClientFactory _httpClientFactory;
-        private readonly ILogger<QbittorrentAdapter> _logger;
-        private readonly QbittorrentAuthSession _authSession;
         private readonly QbittorrentConnectionTester _connectionTester;
+        private readonly QbittorrentAddWorkflow _addWorkflow;
+        private readonly QbittorrentImportMarkerWorkflow _importMarkerWorkflow;
         private readonly QbittorrentRemovalWorkflow _removalWorkflow;
+        private readonly QbittorrentQueueFetchWorkflow _queueFetchWorkflow;
+        private readonly QbittorrentItemFetchWorkflow _itemFetchWorkflow;
         private readonly QbittorrentImportItemResolver _importItemResolver;
 
         public QbittorrentAdapter(IHttpClientFactory httpFactory, ITorrentFileDownloader torrentFileDownloader, ILogger<QbittorrentAdapter> logger)
         {
-            _httpClientFactory = httpFactory ?? throw new ArgumentNullException(nameof(httpFactory));
+            ArgumentNullException.ThrowIfNull(httpFactory);
             _ = torrentFileDownloader ?? throw new ArgumentNullException(nameof(torrentFileDownloader));
-            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
-            _authSession = new QbittorrentAuthSession(_logger);
-            _connectionTester = new QbittorrentConnectionTester(_httpClientFactory, _logger, ClientType);
-            _removalWorkflow = new QbittorrentRemovalWorkflow(_httpClientFactory, _logger, ClientType);
-            _importItemResolver = new QbittorrentImportItemResolver(_logger);
+            ArgumentNullException.ThrowIfNull(logger);
+
+            var authSession = new QbittorrentAuthSession(logger);
+            _connectionTester = new QbittorrentConnectionTester(httpFactory, logger, ClientType);
+            _addWorkflow = new QbittorrentAddWorkflow(httpFactory, authSession, logger, ClientType);
+            _importMarkerWorkflow = new QbittorrentImportMarkerWorkflow(httpFactory, logger, ClientType);
+            _removalWorkflow = new QbittorrentRemovalWorkflow(httpFactory, logger, ClientType);
+            _queueFetchWorkflow = new QbittorrentQueueFetchWorkflow(httpFactory, authSession, logger, ClientType);
+            _itemFetchWorkflow = new QbittorrentItemFetchWorkflow(httpFactory, authSession, logger, ClientType);
+            _importItemResolver = new QbittorrentImportItemResolver(logger);
         }
 
-        public async Task<(bool Success, string Message)> TestConnectionAsync(DownloadClientConfiguration client, CancellationToken ct = default)
-        {
-            return await _connectionTester.TestConnectionAsync(client, ct);
-        }
+        public Task<(bool Success, string Message)> TestConnectionAsync(DownloadClientConfiguration client, CancellationToken ct = default)
+            => _connectionTester.TestConnectionAsync(client, ct);
 
-        public async Task<DownloadClientSubmissionResult> AddAsync(
+        public Task<DownloadClientSubmissionResult> AddAsync(
             DownloadClientConfiguration client,
             PreparedDownloadSubmission submission,
             CancellationToken ct = default)
-        {
-            ArgumentNullException.ThrowIfNull(client);
-            if (submission is not PreparedTorrentSubmission torrent)
-            {
-                throw new DownloadClientSubmissionException("qBittorrent requires a prepared torrent submission.");
-            }
+            => _addWorkflow.AddAsync(client, submission, ct);
 
-            var baseUrl = DownloadClientUriBuilder.BuildAuthority(client);
-            using var httpClient = _httpClientFactory.CreateClient(ClientType);
+        public Task<bool> MarkItemAsImportedAsync(DownloadClientConfiguration client, string downloadId, CancellationToken ct = default)
+            => _importMarkerWorkflow.MarkItemAsImportedAsync(client, downloadId, ct);
 
-            try
-            {
-                await _authSession.LoginAsync(httpClient, client, ct);
-            }
-            catch (QbittorrentException exception)
-            {
-                _logger.LogError(exception, "qBittorrent authentication failed for client {ClientId}", LogRedaction.SanitizeText(client.Id));
-                throw new DownloadClientSubmissionException("qBittorrent authentication failed.", exception);
-            }
+        public Task<bool> RemoveAsync(DownloadClientConfiguration client, string id, bool deleteFiles = false, CancellationToken ct = default)
+            => _removalWorkflow.RemoveAsync(client, id, deleteFiles, ct);
 
-            var addPlan = QbittorrentTorrentAddPlanner.Create(client, torrent);
-
-            using var addContent = QbittorrentAddRequestContentBuilder.Build(addPlan);
-            using var addResponse = await httpClient.PostAsync($"{baseUrl}/api/v2/torrents/add", addContent, ct);
-
-            if (!addResponse.IsSuccessStatusCode)
-            {
-                var responseContent = await addResponse.Content.ReadAsStringAsync(ct);
-                var redacted = LogRedaction.RedactText(responseContent, LogRedaction.GetSensitiveValuesFromEnvironment().Concat([client.Password ?? string.Empty]));
-
-                _logger.LogError($"Failed to add torrent to qBittorrent. Status: {addResponse.StatusCode}, Response: {redacted}");
-                throw new DownloadClientSubmissionException($"qBittorrent rejected the torrent with HTTP {(int)addResponse.StatusCode}.");
-            }
-
-            _logger.LogInformation("Successfully sent torrent to qBittorrent");
-
-            await Task.Delay(1000, ct);
-
-            // Inject tracker URLs via addTrackers API as a fallback to ensure the tracker
-            // is registered even if qBittorrent didn't parse it from the torrent file.
-            if (addPlan.TorrentFileData != null)
-            {
-                try
-                {
-                    var trackerAnnounces = torrent.TrackerUrls.Where(a =>
-                        a.Contains("/announce", StringComparison.OrdinalIgnoreCase) ||
-                        a.Contains("/tracker", StringComparison.OrdinalIgnoreCase)).ToList();
-                    if (trackerAnnounces != null && trackerAnnounces.Count > 0)
-                    {
-                        var trackerUrls = string.Join("\n", trackerAnnounces.Distinct());
-                        using var addTrackersData = new FormUrlEncodedContent(new[]
-                        {
-                            new KeyValuePair<string, string>("hash", addPlan.Hash),
-                            new KeyValuePair<string, string>("urls", trackerUrls)
-                        });
-                        using var trackersResp = await httpClient.PostAsync($"{baseUrl}/api/v2/torrents/addTrackers", addTrackersData, ct);
-                        if (trackersResp.IsSuccessStatusCode)
-                            _logger.LogInformation($"Injected {trackerAnnounces.Count} tracker(s) for torrent {addPlan.Hash} via addTrackers API");
-                        else
-                            _logger.LogDebug($"addTrackers API returned {trackersResp.StatusCode} for torrent {addPlan.Hash} (non-fatal)");
-                    }
-                }
-                catch (Exception exception) when (exception is not (OperationCanceledException or OutOfMemoryException or StackOverflowException))
-                {
-                    _logger.LogDebug(exception, "Non-fatal failure injecting trackers via addTrackers API");
-                }
-            }
-
-            return new DownloadClientSubmissionResult(addPlan.Hash, addPlan.Hash);
-        }
-
-        /// <summary>
-        /// Marks a torrent as imported by changing its category to the configured post-import category.
-        /// This allows users to differentiate imported vs active torrents in qBittorrent.
-        /// Mirrors Sonarr's MarkItemAsImported behavior.
-        /// </summary>
-        public async Task<bool> MarkItemAsImportedAsync(DownloadClientConfiguration client, string downloadId, CancellationToken ct = default)
-        {
-            if (client == null) return false;
-            if (string.IsNullOrEmpty(downloadId)) return false;
-
-            var postImportCategory = client.Settings?.GetValueOrDefault("postImportCategory")?.ToString();
-            if (string.IsNullOrEmpty(postImportCategory))
-            {
-                _logger.LogDebug("No postImportCategory configured for qBittorrent client {ClientId}, skipping MarkItemAsImported", client.Id);
-                return true; // No-op is success
-            }
-
-            var baseUrl = DownloadClientUriBuilder.BuildAuthority(client);
-            try
-            {
-                using var httpClient = _httpClientFactory.CreateClient(ClientType);
-
-                // Authenticate
-                using var loginData = QbittorrentCookieSession.CreateLoginContent(client);
-                using (await httpClient.PostAsync($"{baseUrl}/api/v2/auth/login", loginData, ct)) { }
-
-                // Set category
-                using var setCategoryData = new FormUrlEncodedContent(new[]
-                {
-                    new KeyValuePair<string, string>("hashes", downloadId.ToLowerInvariant()),
-                    new KeyValuePair<string, string>("category", postImportCategory)
-                });
-
-                using var resp = await httpClient.PostAsync($"{baseUrl}/api/v2/torrents/setCategory", setCategoryData, ct);
-                if (resp.IsSuccessStatusCode)
-                {
-                    _logger.LogInformation("Marked torrent {Hash} as imported (category: {Category}) in qBittorrent", downloadId, postImportCategory);
-                    return true;
-                }
-
-                _logger.LogWarning("Failed to mark torrent {Hash} as imported in qBittorrent: {StatusCode}", downloadId, resp.StatusCode);
-                return false;
-            }
-            catch (Exception ex) when (ex is not OperationCanceledException && ex is not OutOfMemoryException && ex is not StackOverflowException)
-            {
-                _logger.LogWarning(ex, "Error marking torrent {Hash} as imported in qBittorrent", downloadId);
-                return false;
-            }
-        }
-
-        public async Task<bool> RemoveAsync(DownloadClientConfiguration client, string id, bool deleteFiles = false, CancellationToken ct = default)
-        {
-            return await _removalWorkflow.RemoveAsync(client, id, deleteFiles, ct);
-        }
-
-        public async Task<List<QueueItem>> GetQueueAsync(DownloadClientConfiguration client, List<string> ids, CancellationToken ct = default)
-        {
-            var items = new List<QueueItem>();
-            if (client == null) return items;
-
-            var isMonitorPoll = ids.Count > 0;
-            var baseUrl = DownloadClientUriBuilder.BuildAuthority(client);
-
-            try
-            {
-                using var httpClient = _httpClientFactory.CreateClient(ClientType);
-                try
-                {
-                    await _authSession.LoginAsync(httpClient, client, ct);
-                }
-                catch (QbittorrentException exception)
-                {
-                    var message = $"qBittorrent authentication failed for client {LogRedaction.SanitizeText(client.Id)}.";
-                    _logger.LogWarning(exception, "qBittorrent authentication failed for client {ClientId}", LogRedaction.SanitizeText(client.Id));
-                    if (isMonitorPoll)
-                    {
-                        throw new DownloadClientAdapterPollingException(message, exception);
-                    }
-                    return items;
-                }
-
-                // Limit fields returned to reduce memory usage
-                var fields = "name,progress,size,downloaded,dlspeed,eta,state,hash,added_on,num_seeds,num_leechs,ratio,save_path";
-
-                // Build category filter parameter if configured
-                var categoryFilter = QBittorrentHelpers.BuildCategoryParameter(client.Settings, "&");
-
-                // Extract category for logging
-                var category = client.Settings?.TryGetValue("category", out var categoryObj) is true
-                    ? categoryObj?.ToString()
-                    : null;
-                QBittorrentHelpers.LogCategoryFiltering(_logger, category);
-
-                // Display requests intentionally fetch a broad snapshot. ID-filtered monitor
-                // requests use qBittorrent's hashes parameter so large queues do not become
-                // unnecessary full snapshots.
-                var hashesFilter = ids.Count > 0
-                    ? $"&hashes={Uri.EscapeDataString(string.Join('|', ids.Select(id => id.ToLowerInvariant())))}"
-                    : string.Empty;
-
-                using var torrentsResp = await httpClient.GetAsync($"{baseUrl}/api/v2/torrents/info?fields={Uri.EscapeDataString(fields)}{categoryFilter}{hashesFilter}", ct);
-                if (!torrentsResp.IsSuccessStatusCode)
-                {
-                    var message = $"qBittorrent queue request failed with status {torrentsResp.StatusCode} for client {LogRedaction.SanitizeText(client.Id)}.";
-                    _logger.LogWarning("qBittorrent queue request failed with status {Status} for client {ClientId}", torrentsResp.StatusCode, LogRedaction.SanitizeText(client.Id));
-                    if (isMonitorPoll)
-                    {
-                        throw new DownloadClientAdapterPollingException(message);
-                    }
-                    return items;
-                }
-
-                var json = await torrentsResp.Content.ReadAsStringAsync(ct);
-                if (string.IsNullOrWhiteSpace(json))
-                {
-                    var message = $"qBittorrent returned an empty queue response for client {LogRedaction.SanitizeText(client.Id)}.";
-                    _logger.LogWarning("qBittorrent returned an empty queue response for client {ClientId}", LogRedaction.SanitizeText(client.Id));
-                    if (isMonitorPoll)
-                    {
-                        throw new DownloadClientAdapterPollingException(message);
-                    }
-                    return items;
-                }
-
-                var torrents = JsonSerializer.Deserialize<List<Dictionary<string, JsonElement>>>(json);
-                if (torrents == null)
-                {
-                    var message = $"qBittorrent returned an invalid queue response for client {LogRedaction.SanitizeText(client.Id)}.";
-                    _logger.LogWarning("qBittorrent returned an invalid queue response for client {ClientId}", LogRedaction.SanitizeText(client.Id));
-                    if (isMonitorPoll)
-                    {
-                        throw new DownloadClientAdapterPollingException(message);
-                    }
-                    return items;
-                }
-
-                foreach (var torrent in torrents)
-                {
-                    var hash = torrent.TryGetValue("hash", out var hashEl) ? hashEl.GetString() ?? string.Empty : string.Empty;
-
-                    List<Dictionary<string, JsonElement>> files = [];
-                    using var filesResp = await httpClient.GetAsync($"{baseUrl}/api/v2/torrents/files?hash={hash}", ct);
-                    if (filesResp.IsSuccessStatusCode)
-                    {
-                        var filesJson = await filesResp.Content.ReadAsStringAsync(ct);
-                        files = JsonSerializer.Deserialize<List<Dictionary<string, JsonElement>>>(filesJson) ?? [];
-                    }
-
-                    items.Add(QbittorrentResponseMapper.MapQueueItem(torrent, client, files));
-                }
-            }
-            catch (DownloadClientAdapterPollingException)
-            {
-                throw;
-            }
-            catch (Exception ex) when (ex is not OperationCanceledException && ex is not OutOfMemoryException && ex is not StackOverflowException)
-            {
-                _logger.LogWarning(ex, "Error getting qBittorrent queue - client may be unreachable");
-                if (isMonitorPoll)
-                {
-                    throw new DownloadClientAdapterPollingException("Error polling qBittorrent queue.", ex);
-                }
-            }
-
-            return FilterByIds(items, ids);
-        }
+        public Task<List<QueueItem>> GetQueueAsync(DownloadClientConfiguration client, List<string> ids, CancellationToken ct = default)
+            => _queueFetchWorkflow.GetQueueAsync(client, ids, ct);
 
         public Task<List<QueueItem>> GetQueueAsync(DownloadClientConfiguration client, CancellationToken ct = default)
             => GetQueueAsync(client, [], ct);
 
         public Task<List<(string Id, string Name)>> GetRecentHistoryAsync(DownloadClientConfiguration client, int limit = 100, CancellationToken ct = default)
-        {
-            return Task.FromResult(new List<(string Id, string Name)>());
-        }
+            => Task.FromResult(new List<(string Id, string Name)>());
+
+        public Task<List<DownloadClientItem>> GetItemsAsync(DownloadClientConfiguration client, CancellationToken ct = default)
+            => _itemFetchWorkflow.GetItemsAsync(client, ct);
 
         /// <summary>
-        /// Get all downloads as standardized DownloadClientItem objects
+        /// Get import item from DownloadClientItem.
         /// </summary>
-        public async Task<List<DownloadClientItem>> GetItemsAsync(DownloadClientConfiguration client, CancellationToken ct = default)
-        {
-            var items = new List<DownloadClientItem>();
-            if (client == null) return items;
-
-            var baseUrl = DownloadClientUriBuilder.BuildAuthority(client);
-            var categoryFilter = QBittorrentHelpers.BuildCategoryParameter(client.Settings, "&");
-
-            try
-            {
-                using var httpClient = QbittorrentCookieSession.CreateClient();
-                using var loginData = QbittorrentCookieSession.CreateLoginContent(client);
-
-                using var loginResp = await httpClient.PostAsync($"{baseUrl}/api/v2/auth/login", loginData, ct);
-                if (loginResp.StatusCode == HttpStatusCode.Forbidden)
-                {
-                    using var testResp = await httpClient.GetAsync($"{baseUrl}/api/v2/app/version", ct);
-                    if (!testResp.IsSuccessStatusCode)
-                    {
-                        _logger.LogWarning("qBittorrent authentication appears to be enabled and credentials are invalid for client {ClientId}", LogRedaction.SanitizeText(client.Id));
-                        return items;
-                    }
-                }
-                else if (!loginResp.IsSuccessStatusCode)
-                {
-                    _logger.LogWarning("qBittorrent login failed with status {Status} for client {ClientId}", loginResp.StatusCode, LogRedaction.SanitizeText(client.Id));
-                    return items;
-                }
-
-                // Fetch qBittorrent global preferences for seed limit evaluation (Sonarr parity)
-                bool globalMaxRatioEnabled = false;
-                float globalMaxRatio = -1f;
-                bool globalMaxSeedingTimeEnabled = false;
-                long globalMaxSeedingTime = -1;
-                try
-                {
-                    using var prefsResp = await httpClient.GetAsync($"{baseUrl}/api/v2/app/preferences", ct);
-                    if (prefsResp.IsSuccessStatusCode)
-                    {
-                        var prefsJson = await prefsResp.Content.ReadAsStringAsync(ct);
-                        if (!string.IsNullOrWhiteSpace(prefsJson))
-                        {
-                            var prefs = JsonSerializer.Deserialize<Dictionary<string, JsonElement>>(prefsJson);
-                            if (prefs != null)
-                            {
-                                globalMaxRatioEnabled = prefs.TryGetValue("max_ratio_enabled", out var mre) && mre.GetBoolean();
-                                globalMaxRatio = prefs.TryGetValue("max_ratio", out var mr) ? (float)mr.GetDouble() : -1f;
-                                globalMaxSeedingTimeEnabled = prefs.TryGetValue("max_seeding_time_enabled", out var mste) && mste.GetBoolean();
-                                globalMaxSeedingTime = prefs.TryGetValue("max_seeding_time", out var mst) ? mst.GetInt64() : -1;
-                            }
-                        }
-                    }
-                }
-                catch (Exception ex) when (ex is not OperationCanceledException && ex is not OutOfMemoryException && ex is not StackOverflowException)
-                {
-                    _logger.LogDebug(ex, "Failed to fetch qBittorrent preferences for seed limit evaluation, will use conservative defaults");
-                }
-
-                // Resolve removeCompletedDownloads setting once for all torrents
-                var removeCompletedDownloads = !string.IsNullOrEmpty(client.RemoveCompletedDownloads) &&
-                    client.RemoveCompletedDownloads != "none";
-
-                // Limit fields returned to reduce memory usage
-                var fields = "name,progress,size,downloaded,dlspeed,eta,state,hash,added_on,num_seeds,num_leechs,ratio,save_path,category,content_path,ratio_limit,seeding_time_limit,seeding_time";
-                using var torrentsResp = await httpClient.GetAsync($"{baseUrl}/api/v2/torrents/info?fields={Uri.EscapeDataString(fields)}{categoryFilter}", ct);
-                if (!torrentsResp.IsSuccessStatusCode) return items;
-
-                var json = await torrentsResp.Content.ReadAsStringAsync(ct);
-                if (string.IsNullOrWhiteSpace(json)) return items;
-
-                var torrents = JsonSerializer.Deserialize<List<Dictionary<string, JsonElement>>>(json);
-                if (torrents == null) return items;
-
-                foreach (var torrent in torrents)
-                {
-                    items.Add(QbittorrentResponseMapper.MapDownloadClientItem(
-                        torrent,
-                        client,
-                        removeCompletedDownloads,
-                        globalMaxRatioEnabled,
-                        globalMaxRatio,
-                        globalMaxSeedingTimeEnabled,
-                        globalMaxSeedingTime));
-                }
-            }
-            catch (Exception ex) when (ex is not OperationCanceledException && ex is not OutOfMemoryException && ex is not StackOverflowException)
-            {
-                _logger.LogWarning(ex, "Error getting qBittorrent items - client may be unreachable");
-            }
-
-            return items;
-        }
-
-        /// <summary>
-        /// Get import item from DownloadClientItem
-        /// </summary>
-        public async Task<DownloadClientItem> GetImportItemAsync(
+        public Task<DownloadClientItem> GetImportItemAsync(
             DownloadClientConfiguration client,
             DownloadClientItem item,
             DownloadClientItem? previousAttempt = null,
             CancellationToken ct = default)
-        {
-            return await _importItemResolver.GetImportItemAsync(client, item, ct);
-        }
+            => _importItemResolver.GetImportItemAsync(client, item, ct);
 
         /// <summary>
         /// LEGACY: Resolves the actual import item for a completed download.
         /// Matches GetImportItem pattern.
         /// </summary>
-        public async Task<QueueItem> GetImportItemAsync(
+        public Task<QueueItem> GetImportItemAsync(
             DownloadClientConfiguration client,
             Download download,
             QueueItem queueItem,
             QueueItem? previousAttempt = null,
             CancellationToken ct = default)
-        {
-            return await _importItemResolver.GetImportItemAsync(client, download, queueItem, ct);
-        }
+            => _importItemResolver.GetImportItemAsync(client, download, queueItem, ct);
 
         internal static string ResolveTorrentContentPath(
             string savePath,
             List<Dictionary<string, JsonElement>> files)
         {
             return QbittorrentImportPathResolver.ResolveContentPath(savePath, files);
-        }
-
-        private static List<QueueItem> FilterByIds(List<QueueItem> items, List<string> ids)
-        {
-            if (ids.Count == 0)
-            {
-                return items;
-            }
-
-            var idSet = ids.ToHashSet(StringComparer.OrdinalIgnoreCase);
-            return [.. items.Where(item => !string.IsNullOrWhiteSpace(item.Id) && idSet.Contains(item.Id))];
         }
     }
 }
